requirements.txt
pandas==2.0.3
sentence-transformers==2.6.1
langchain==0.1.16